requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.4
django-ckeditor==6.7.0
django-cuser==2017.3.16
django-js-asset==2.1.0
django-modeltranslation==0.18.11
django-rosetta==0.9.9
django-tinymce==3.6.1
idna==3.4
oauthlib==3.2.2
Pillow==10.0.0
polib==1.2.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
